pyproject.toml
[project]
name = "stego-side-wing"
version = "0.1.0"
description = "Python backend and workflow runtime for stego pipelines (API, workflows, metrics)."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.13.3",
    "beautifulsoup4>=4.14.3",
    "crawl4ai>=0.8.0",
    "flask-caching>=2.3.1",
    "flask-cors>=6.0.1",
    "flask[async]>=3.1.2",
    "google-genai>=1.0.0",
    "httpx>=0.28.1",
    "icecream>=2.1.10",
    "langdetect>=1.0.9",
    "loguru>=0.7.3",
    "ollama>=0.6.1",
    "openai>=2.26.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "selenium>=4.41.0",
    "sentence-transformers>=5.3.0",
]

[project.optional-dependencies]
metrics = [
    "torch>=2.6.0",
    "transformers>=4.50.0",
]

[dependency-groups]
dev = [
    "pyright>=1.1.408",
    "pytest>=9.0.2",
    "requests-toolbelt>=1.0.0",
    "ruff>=0.9.0",
]

[tool.pytest.ini_options]
testpaths = ["src/tests"]
addopts = "-q"
filterwarnings = [
    "ignore::requests.exceptions.RequestsDependencyWarning",
]

[tool.ruff]
target-version = "py313"
src = ["src"]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
